requirements.txt
attrs==25.3.0
audioread==3.0.1
automat==25.4.16
causal_conv1d==1.5.0.post8
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
greenlet==3.2.1
grpcio==1.71.0
grpcio-tools==1.71.0
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.4.0
mamba_ssm==2.2.4
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
ninja==1.11.1.4
noisereduce==3.0.3
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
parsel==1.10.0
Pillow==11.2.1
pip==25.1
platformdirs==4.3.7
playwright==1.51.0
pooch==1.8.2
protobuf==5.29.4
protego==0.4.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
pyparsing==3.2.3
PyDispatcher==2.0.7
pyee==12.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
queuelib==1.8.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
scrapy-playwright==0.0.43
service-identity==24.2.0
setuptools==79.0.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.3
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
timm==1.0.15
tldextract==5.3.0
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
Twisted==24.11.0
typing_extensions==4.13.2
urllib3==2.4.0
w3lib==2.3.1
zope.interface==7.2